UINT -> uint32_t

+++ src/Modules/Context/Context.h
@@ -20,20 +20,20 @@ public:
 
     virtual std::unique_ptr<ProgramApi> CreateProgram() = 0;
     virtual Resource::Ptr CreateTexture(uint32_t bind_flag, DXGI_FORMAT format, uint32_t msaa_count, int width, int height, int depth = 1, int mip_levels = 1) = 0;
-    virtual Resource::Ptr CreateBuffer(uint32_t bind_flag, UINT buffer_size, size_t stride) = 0;
-    virtual void UpdateSubresource(const Resource::Ptr& ires, UINT DstSubresource, const void *pSrcData, UINT SrcRowPitch, UINT SrcDepthPitch) = 0;
+    virtual Resource::Ptr CreateBuffer(uint32_t bind_flag, uint32_t buffer_size, uint32_t stride) = 0;
+    virtual void UpdateSubresource(const Resource::Ptr& ires, uint32_t DstSubresource, const void *pSrcData, uint32_t SrcRowPitch, uint32_t SrcDepthPitch) = 0;
 
-    virtual void SetViewport(int width, int height) = 0;
-    virtual void SetScissorRect(LONG left, LONG top, LONG right, LONG bottom) = 0;
+    virtual void SetViewport(float width, float height) = 0;
+    virtual void SetScissorRect(int32_t left, int32_t top, int32_t right, int32_t bottom) = 0;
 
-    virtual void IASetIndexBuffer(Resource::Ptr res, UINT SizeInBytes, DXGI_FORMAT Format) = 0;
-    virtual void IASetVertexBuffer(UINT slot, Resource::Ptr res, UINT SizeInBytes, UINT Stride) = 0;
+    virtual void IASetIndexBuffer(Resource::Ptr res, uint32_t SizeInBytes, DXGI_FORMAT Format) = 0;
+    virtual void IASetVertexBuffer(uint32_t slot, Resource::Ptr res, uint32_t SizeInBytes, uint32_t Stride) = 0;
 
     virtual void BeginEvent(LPCWSTR Name) = 0;
     virtual void EndEvent() = 0;
 
-    virtual void DrawIndexed(UINT IndexCount, UINT StartIndexLocation, INT BaseVertexLocation) = 0;
-    virtual void Dispatch(UINT ThreadGroupCountX, UINT ThreadGroupCountY, UINT ThreadGroupCountZ) = 0;
+    virtual void DrawIndexed(uint32_t IndexCount, uint32_t StartIndexLocation, int32_t BaseVertexLocation) = 0;
+    virtual void Dispatch(uint32_t ThreadGroupCountX, uint32_t ThreadGroupCountY, uint32_t ThreadGroupCountZ) = 0;
 
     virtual Resource::Ptr GetBackBuffer() = 0;
     virtual void Present(const Resource::Ptr& ires) = 0;
