+++ src/Modules/Context/DX11Context.h
@@ -18,20 +18,20 @@ public:
 
     virtual std::unique_ptr<ProgramApi> CreateProgram() override;
     virtual Resource::Ptr CreateTexture(uint32_t bind_flag, DXGI_FORMAT format, uint32_t msaa_count, int width, int height, int depth = 1, int mip_levels = 1) override;
-    virtual Resource::Ptr CreateBuffer(uint32_t bind_flag, UINT buffer_size, size_t stride) override;
-    virtual void UpdateSubresource(const Resource::Ptr& ires, UINT DstSubresource, const void *pSrcData, UINT SrcRowPitch, UINT SrcDepthPitch) override;
+    virtual Resource::Ptr CreateBuffer(uint32_t bind_flag, uint32_t buffer_size, uint32_t stride) override;
+    virtual void UpdateSubresource(const Resource::Ptr& ires, uint32_t DstSubresource, const void *pSrcData, uint32_t SrcRowPitch, uint32_t SrcDepthPitch) override;
 
-    virtual void SetViewport(int width, int height) override;
-    virtual void SetScissorRect(LONG left, LONG top, LONG right, LONG bottom) override;
+    virtual void SetViewport(float width, float height) override;
+    virtual void SetScissorRect(int32_t left, int32_t top, int32_t right, int32_t bottom) override;
 
-    virtual void IASetIndexBuffer(Resource::Ptr res, UINT SizeInBytes, DXGI_FORMAT Format) override;
-    virtual void IASetVertexBuffer(UINT slot, Resource::Ptr res, UINT SizeInBytes, UINT Stride) override;
+    virtual void IASetIndexBuffer(Resource::Ptr res, uint32_t SizeInBytes, DXGI_FORMAT Format) override;
+    virtual void IASetVertexBuffer(uint32_t slot, Resource::Ptr res, uint32_t SizeInBytes, uint32_t Stride) override;
 
     virtual void BeginEvent(LPCWSTR Name) override;
     virtual void EndEvent() override;
 
-    virtual void DrawIndexed(UINT IndexCount, UINT StartIndexLocation, INT BaseVertexLocation) override;
-    virtual void Dispatch(UINT ThreadGroupCountX, UINT ThreadGroupCountY, UINT ThreadGroupCountZ) override;
+    virtual void DrawIndexed(uint32_t IndexCount, uint32_t StartIndexLocation, INT BaseVertexLocation) override;
+    virtual void Dispatch(uint32_t ThreadGroupCountX, uint32_t ThreadGroupCountY, uint32_t ThreadGroupCountZ) override;
 
     virtual Resource::Ptr GetBackBuffer() override;
     virtual void Present(const Resource::Ptr& ires) override;
